requirements.txt
python-dotenv
PyMySQL
mysql-connector-python
beautifulsoup4
requests
pyTelegramBotAPI